+++ public/index2.html
@@ -1,0 +1,64 @@
+<!doctype html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <link rel="stylesheet" href="style/style.css" />
+    <link rel="stylesheet" href="style/pages/landing.css" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <meta
+      name="description"
+      content="세상의 모든 정보를 쉽게 저장하고 관리해 보세요"
+    />
+    <meta property="og:title" content="Linkbrary" />
+    <meta property="og:url" content="https://youdame.netlify.app/" />
+    <meta property="og:type" content="website" />
+    <meta
+      property="og:description"
+      content="세상의 모든 정보를 쉽게 저장하고 관리해 보세요"
+    />
+    <meta
+      property="og:image"
+      content="https://youdame.netlify.app/images/landing/img1.png"
+    />
+
+    <title>Linkbrary</title>
+  </head>
+
+  <body>
+    <header>
+      <div class="starter">
+        <a class="logo" href="/">
+          <img src="images/landing/logo.png" />
+        </a>
+        <a class="login flex-center" href="signin.html">로그인</a>
+      </div>
+    </header>
+
+    <main></main>
+    <footer>
+      <div class="frame">
+        <div class="footer-container">
+          <div class="codeit-2023">code-it - 2023</div>
+          <div class="extra-info">
+            <a href="privacy.html"><span>Privacy Policy</span></a>
+            <a href="faq.html"><span>FAQ</span></a>
+          </div>
+          <div class="sns-icons">
+            <a href="https://www.facebook.com"
+              ><img src="images/landing/facebook.png"
+            /></a>
+            <a href="https://twitter.com"
+              ><img src="images/landing/twitter.png"
+            /></a>
+            <a href="https://www.youtube.com"
+              ><img src="images/landing/youtube.png"
+            /></a>
+            <a href="https://www.instagram.com"
+              ><img src="images/landing/instagram.png"
+            /></a>
+          </div>
+        </div>
+      </div>
+    </footer>
+  </body>
+</html>
