+++ src/style/landing.css
@@ -1,0 +1,401 @@
+* {
+  box-sizing: border-box;
+}
+
+html, body, div, span, h1, h2, h3, h4, h5, h6, p, pre {
+  word-break: keep-all;
+}
+
+body {
+  margin: 0 auto;
+}
+
+header {
+  height: 5.875rem;
+  display: flex;
+  padding: 1.25rem 12.5rem;
+  flex-direction: column;
+  align-items: center;
+  gap: 0.5rem;
+  align-self: stretch;
+  background-color: var(--linkbrary-bg);
+  position: sticky;
+  top: 0;
+}
+
+header .starter {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  align-self: stretch;
+}
+
+header .starter a,
+header .starter button,
+main .intro .addlink,
+footer .frame .footer-container .extra-info a,
+footer .frame .footer-container .sns-icons img {
+  cursor: pointer;
+}
+
+header .login {
+  border: none;
+  color: var(--grey-light);
+  font-family: Pretendard;
+  font-size: 1.125rem;
+  font-weight: 600;
+
+  width: 128px;
+  height: 54px;
+  padding: 1rem 1.25rem;
+
+  gap: 0.625rem;
+  border-radius: 8px;
+  background: var(--gra-purpleblue-to-skyblue);
+  text-decoration-line: none;
+}
+
+main .intro {
+  display: flex;
+  padding: 70px 360px 0px 360px;
+  flex-direction: column;
+  justify-content: flex-end;
+  background: #f0f6ff;
+  align-items: center;
+  gap: 2.5rem;
+}
+
+main .intro h1 {
+  width: 44rem;
+  display: flex;
+  flex-direction: column;
+  text-align: center;
+  font-family: Pretendard;
+  font-size: 3.6rem;
+  font-weight: 700;
+  line-height: 80px;
+}
+
+main .intro h1 .second-line > * {
+  display: inline;
+}
+main .intro h1 span {
+  background: linear-gradient(91deg, #6d6afe 17.28%, #ff9f9f 74.98%);
+  background-clip: text;
+  -webkit-background-clip: text;
+  -webkit-text-fill-color: transparent;
+}
+
+main .intro .add-link {
+  width: 21.875rem;
+  padding: 1rem 1.25rem;
+  gap: 0.625rem;
+  border-radius: 8px;
+  background: var(--gra-purpleblue-to-skyblue);
+  color: var(--grey-light);
+  font-family: Pretendard;
+  font-size: 1.125rem;
+  font-weight: 600;
+  border: none;
+  text-decoration-line: none;
+}
+
+main .intro .intro-img {
+  width: 1100px;
+  height: 659px;
+  flex-shrink: 0;
+  border-radius: 25px;
+  background: linear-gradient(
+    180deg,
+    rgba(196, 196, 196, 0) 67.68%,
+    #f0f6ff 94.76%
+  );
+}
+
+main .content {
+  padding: 7.5rem 28.8125rem 3.125rem 28.8125rem;
+}
+
+.content .frame {
+  display: grid;
+  grid-template-rows: 1fr 1fr;
+
+  grid-auto-columns: auto;
+  justify-content: center;
+  align-items: center;
+
+  gap: 0 9em;
+}
+
+.frame .texts {
+  display: flex;
+  width: 18rem;
+  flex-direction: column;
+  align-items: flex-start;
+  gap: 0.625rem;
+  line-height: normal;
+  grid-row: 1;
+  grid-column: 1;
+}
+
+.frame .texts h2 {
+  font-family: Pretendard;
+  font-size: 3rem;
+  font-weight: 700;
+  letter-spacing: -0.3px;
+  padding-top: 140px;
+}
+
+.frame .texts.reloc-texts {
+  grid-row: 1;
+  grid-column: 2;
+}
+
+.frame .texts.pink-blur {
+  background: linear-gradient(96deg, #fe8a8a 1.72%, #a4ceff 74.97%);
+  background-clip: text;
+  -webkit-background-clip: text;
+}
+
+.frame .texts.beige-blur {
+  background: linear-gradient(277deg, #6fbaff 59.22%, #ffd88b 93.66%);
+  background-clip: text;
+  -webkit-background-clip: text;
+}
+
+.frame .texts.blue-blur {
+  background: linear-gradient(
+    99deg,
+    #6d7ccd 19.76%,
+    rgba(82, 136, 133, 0.22) 52.69%
+  );
+  background-clip: text;
+  -webkit-background-clip: text;
+}
+
+.frame .texts.sky-blur {
+  background: var(--fandom-k-gra-blue-to-pink);
+  background-clip: text;
+  -webkit-background-clip: text;
+}
+
+.frame .texts h2 span {
+  -webkit-text-fill-color: transparent;
+}
+
+.frame .para {
+  color: #6b6b6b;
+  font-family: Pretendard;
+  font-size: 1rem;
+  font-weight: 500;
+  line-height: 150%;
+  padding-bottom: 160px;
+}
+
+.frame .para.reloc-para {
+  grid-row: 2;
+  grid-column: 2;
+}
+
+.content .frame .content-img {
+  width: 550px;
+  height: 450px;
+  border-radius: 15px;
+  background: #f0f6ff;
+  grid-row: 1 / span 2;
+  grid-column: 2;
+}
+
+.content .frame .content-img.reloc-img {
+  grid-row: 1 / span 2;
+  grid-column: 1;
+}
+
+footer {
+  display: flex;
+  height: 17.5rem;
+  padding-top: 120px;
+  flex-direction: column;
+  align-items: center;
+  align-self: stretch;
+}
+
+footer .frame {
+  display: flex;
+  height: 10rem;
+  padding: 2rem 6.5rem 4rem 6.5rem;
+  flex-direction: column;
+  align-items: center;
+  flex-shrink: 0;
+  align-self: stretch;
+  background: var(--the-julge-black);
+}
+
+footer .frame .footer-container {
+  display: flex;
+  justify-content: space-between;
+  align-items: flex-start;
+  align-self: stretch;
+  font-family: Arial;
+  font-size: 1rem;
+  font-weight: 400;
+}
+
+footer .frame .footer-container .codeit-2023 {
+  color: #676767;
+  text-align: center;
+}
+
+footer .frame .footer-container .extra-info {
+  display: flex;
+  width: 11rem;
+  align-items: flex-start;
+  gap: 1.875rem;
+  justify-content: center;
+}
+
+footer .frame .footer-container .extra-info a {
+  color: #cfcfcf;
+  text-decoration: none;
+}
+
+footer .frame .footer-container .sns-icons {
+  display: flex;
+  align-items: flex-start;
+  gap: 0.75rem;
+}
+
+footer .frame .footer-container .sns-icons img {
+  width: 1.25rem;
+  height: 1.25rem;
+}
+
+/* 태블릿 화면 크기 (768px 이상 ~ 1199px 이하)에 대한 스타일 */
+@media (min-width: 768px) and (max-width: 1199px) {
+  header {
+    padding: 1.25rem 32px;
+  }
+
+  header .logo {
+    margin-left: auto;
+    margin-right: 480px;
+  }
+
+  header .login {
+    margin-right: auto;
+  }
+
+  .content .frame .content-img {
+    width: 330px;
+    height: 315px;
+  }
+
+  main .content {
+    padding: 7.5rem 23rem 3.125rem 23rem;
+  }
+
+  main .intro .intro-img {
+    width: 698px;
+    height: 343px;
+  }
+
+  main .intro h1 .second-line > * {
+    display: block;
+  }
+}
+
+@media (max-width: 767px) {
+  header,
+  main .intro,
+  .content.flex-center,
+  footer .frame {
+    width: 100%;
+    padding-left: 32px;
+    padding-right: 32px;
+  }
+
+  main .intro h1 .second-line > * {
+    display: block;
+  }
+  .content.flex-center {
+    padding-top: 40px;
+    padding-bottom: 40px;
+  }
+
+  .content.flex-center.first-content {
+    padding-top: 80px;
+  }
+  main .intro .intro-img,
+  .content .frame,
+  .content .frame .texts {
+    width: 100%;
+    height: 100%;
+  }
+
+  .content .frame .content-img {
+    width: 100%;
+    height: auto;
+  }
+
+  .content .frame {
+    display: grid;
+    grid-template-rows: 0.2fr 1fr 0.2fr;
+    grid-auto-rows: 1fr;
+    gap: 30px;
+  }
+
+  .frame .texts h2 {
+    padding-top: 0;
+  }
+
+  .frame .para {
+    padding-bottom: 0;
+  }
+
+  .content .frame .texts h2,
+  .content .frame .reloc-texts {
+    grid-row: 1;
+    grid-column: 1;
+  }
+  .content .frame .content-img,
+  .content .frame .content-img.reloc-img {
+    grid-row: 2;
+    grid-column: 1;
+  }
+
+  .content .frame .texts .para,
+  .content .frame .para.reloc-para {
+    grid-row: 3;
+    grid-column: 1;
+  }
+
+  footer .frame .footer-container {
+    display: grid;
+    grid-template-columns: 1fr 1fr;
+    grid-auto-rows: auto;
+    gap: 3em;
+  }
+
+  footer .frame .codeit-2023 {
+    grid-row: 2;
+    grid-column: 1;
+  }
+
+  footer .frame .extra-info {
+    grid-row: 1;
+    grid-column: 1;
+  }
+
+  footer .frame .footer-container .sns-icons {
+    grid-row: 1;
+    grid-column: 2;
+    justify-content: flex-end;
+  }
+
+  footer .frame .footer-container .codeit-2023 {
+    text-align: start;
+  }
+  footer .frame .footer-container .extra-info {
+    justify-content: flex-start;
+  }
+}
